(#1406) Move Callable classes to `org.cactoos.callables`
